posts.component moved to posts-container directory
